[Refactor] Extract networking types
