+++ src/utils/message-for.ts
@@ -1,8 +1,9 @@
 import { hasComponent } from "bitecs";
 import { HubsWorld } from "../app";
 import { Networked } from "../bit-components";
-import { EntityID, Message, createMessageDatas, UpdateMessage } from "../bit-systems/networking";
+import { createMessageDatas } from "../bit-systems/networking";
 import { networkableComponents, schemas } from "./network-schemas";
+import type { EntityID, Message, UpdateMessage } from "./networking-types";
 
 export function messageFor(
   world: HubsWorld,
