+++ src/utils/take-ownership.ts
@@ -2,7 +2,7 @@ import { addComponent, hasComponent } from "bitecs";
 import { HubsWorld } from "../app";
 import { AEntity, Networked, Owned } from "../bit-components";
 import { getServerTime } from "../phoenix-adapter";
-import { EntityID } from "../bit-systems/networking";
+import type { EntityID } from "./networking-types";
 
 export function takeOwnership(world: HubsWorld, eid: EntityID) {
   // TODO we do this to have a single API for taking ownership of things in new code, but it obviously relies on NAF/AFrame
